requirements-dev.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django>=2.2.4
django-autocomplete-light==3.3.5
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
django-filter==2.1.0
django-guardian==2.0.0
django-mathfilters==0.4.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.9.4
django-tables2==2.1.1
django-queryset-csv==1.0.2
idna==2.8
itsdangerous==1.1.0
Markdown==3.1.1
pycparser==2.19
pytz==2019.1
psycopg2
requests==2.22.0
six==1.12.0
snowflake==0.0.3
sqlparse==0.3.0
urllib3==1.25.3
iso8601==0.1.12
